refac: Move Docs pages into its dir
