+++ src/Pages/Docs/Pages/BatchAutionsOverview.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material'
-import './index.scss'
+import '../index.scss'
 
 const BatchAutionsOverview: React.FC = () => {
   return (
