+++ src/Pages/Docs/Pages/Benefits.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material'
-import './index.scss'
+import '../index.scss'
 
 const Benefits: React.FC = () => {
   return (
