+++ src/Pages/Docs/Pages/FairProtocolOverview.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material'
-import './index.scss'
+import '../index.scss'
 
 const FairProtocolOverview: React.FC = () => {
   return (
