+++ src/Pages/Docs/Pages/Features.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material'
-import './index.scss'
+import '../index.scss'
 
 const Features: React.FC = () => {
   return (
