+++ src/Pages/Docs/Pages/Github.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material'
-import './index.scss'
+import '../index.scss'
 
 const Github: React.FC = () => {
   return (
