+++ src/Pages/Docs/Pages/ParticipatingRequirements.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material'
-import './index.scss'
+import '../index.scss'
 
 const ParticipatingRequirements: React.FC = () => {
   return (
