+++ src/Pages/Docs/Pages/SettlingAuctions.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material'
-import './index.scss'
+import '../index.scss'
 
 const SettlingAuctions: React.FC = () => {
   return (
