+++ src/Pages/Docs/Pages/UseCases.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material'
-import './index.scss'
+import '../index.scss'
 
 const UseCases: React.FC = () => {
   return (
