+++ src/Pages/Docs/Pages/VestedTokens.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material'
-import './index.scss'
+import '../index.scss'
 
 const VestedTokens: React.FC = () => {
   return (
