+++ src/Pages/Docs/index.tsx
@@ -25,22 +25,22 @@ import Image9 from 'src/assets/images/image9.png'
 
 import { DocsDetail, DocsItem, docsDetails, docsItems } from './utils'
 import './index.scss'
-import UserFlow from './UserFlow'
-import FairProtocolOverview from './FairProtocolOverview'
-import Features from './Features'
-import SupportedNetworks from './SupportedNetworks'
-import BatchAutionsOverview from './BatchAutionsOverview'
-import Benefits from './Benefits'
-import UseCases from './UseCases'
-import VestedTokens from './VestedTokens'
-import ParticipatingRequirements from './ParticipatingRequirements'
-import StartingRequirements from './StartingRequirements'
-import UsingScripts from './UsingScripts'
-import UsingUI from './UsingUI'
-import SettlingAuctions from './SettlingAuctions'
-import Github from './Github'
-import PrivateAuctions from './PrivateAuctions'
-import Faq from './Faq'
+import UserFlow from './Pages/UserFlow'
+import FairProtocolOverview from './Pages/FairProtocolOverview'
+import Features from './Pages/Features'
+import SupportedNetworks from './Pages/SupportedNetworks'
+import BatchAutionsOverview from './Pages/BatchAutionsOverview'
+import Benefits from './Pages/Benefits'
+import UseCases from './Pages/UseCases'
+import VestedTokens from './Pages/VestedTokens'
+import ParticipatingRequirements from './Pages/ParticipatingRequirements'
+import StartingRequirements from './Pages/StartingRequirements'
+import UsingScripts from './Pages/UsingScripts'
+import UsingUI from './Pages/UsingUI'
+import SettlingAuctions from './Pages/SettlingAuctions'
+import Github from './Pages/Github'
+import PrivateAuctions from './Pages/PrivateAuctions'
+import Faq from './Pages/Faq'
 
 const Docs: React.FC = () => {
   const isMobileOrTablet = useMediaQuery('(max-width:960px)')
